\name{matrix.vector}
\alias{rows}
\alias{cols}
\alias{dim.all}
\alias{concatenate}

\title{The functins allow to manipulate vectors as matrices}
\description{
 By default, Vectors are considered as matrices with one column
}
\usage{
rows(a, as.row = FALSE)
cols(a, as.row = FALSE)
dim.all(a, as.row = FALSE)
concatenate(edges.hist, vec.mat, constant = NULL)
}

\keyword{internal}
