\name{network}
\alias{local.network}
\alias{global.network}

\title{creates the network based in the edge distribution and the number of nodes for local networks}
\description{
  Internal function
}
\usage{
local.network(n, distrib, param, one.connection = TRUE)
global.network(n, p, distrib, param = 0, distrib2 = "full", param2 = 0)
}

\keyword{internal}
