\name{simulation}
\alias{epidemic.sim}
\alias{stime.infection}

\title{ epidemic simulation }
\description{
  Internal function called by "epidemic"
}
\usage{
epidemic.sim(n, p, obs.time = 0, ini.inf, distrib, param, distrib2, param2, BETA, GAMMA, ETA, SUSC)
stime.infection(network.edges, indexes, infectives, BETA)
}


\keyword{internal}
