\name{sonia.formatting}
\alias{sformat.output}
\alias{angles}
\alias{coordinates}
\title{ These functions prepare the epidemic.sim output as a SoNIA input }
\description{
  Internal function
}
\usage{
sformat.output(a, random = FALSE)
angles(X)
coordinates(n, indexes, random = FALSE)
}

\keyword{internal}
