\name{tasks}
\alias{grouping}
\alias{elements.matrix}
\alias{vector.one}
\alias{order.edges}
\alias{resample}
\alias{number.elements}
\alias{compare.vectors}

\title{ Minor functions}
\description{
Internal function
}
\usage{
grouping(vec.mat, groups)
elements.matrix(matrix, esc, cases = 0, rows.mat = 0, as.row = FALSE)
vector.one(n, q)
order.edges(edges, ord.col = TRUE)
resample(x, size, rep = FALSE)
number.elements(mat, col, x)
compare.vectors(y, x)
}
\keyword{internal}
